requirements.txt
colorzero==2.0
gpiozero==1.6.2
numpy==1.24.2
pigpio==1.78
RPi.GPIO==0.7.1
